requirements.txt
azure-functions==1.17.0
azure-identity==1.15.0
requests==2.31.0